+++ benchmark/competition/array_of_structs/array_of_pointers_test.go
@@ -14,7 +14,9 @@ func runAoP16B(b *testing.B, count int) {
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		for _, e := range entities {
+		count := len(entities)
+		for i := 0; i < count; i++ {
+			e := entities[i]
 			e.S0.Hi++
 			e.S0.Lo++
 		}
@@ -31,7 +33,9 @@ func runAoP32B(b *testing.B, count int) {
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		for _, e := range entities {
+		count := len(entities)
+		for i := 0; i < count; i++ {
+			e := entities[i]
 			e.S0.Hi++
 			e.S0.Lo++
 		}
@@ -48,7 +52,9 @@ func runAoP64B(b *testing.B, count int) {
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		for _, e := range entities {
+		count := len(entities)
+		for i := 0; i < count; i++ {
+			e := entities[i]
 			e.S0.Hi++
 			e.S0.Lo++
 		}
@@ -65,7 +71,9 @@ func runAoP128B(b *testing.B, count int) {
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		for _, e := range entities {
+		count := len(entities)
+		for i := 0; i < count; i++ {
+			e := entities[i]
 			e.S0.Hi++
 			e.S0.Lo++
 		}
@@ -82,7 +90,9 @@ func runAoP256B(b *testing.B, count int) {
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		for _, e := range entities {
+		count := len(entities)
+		for i := 0; i < count; i++ {
+			e := entities[i]
 			e.S0.Hi++
 			e.S0.Lo++
 		}
@@ -101,10 +111,6 @@ func BenchmarkArrOfPointers_16B_100_000(b *testing.B) {
 	runAoP16B(b, 100000)
 }
 
-func BenchmarkArrOfPointers_16B_250_000(b *testing.B) {
-	runAoP16B(b, 250000)
-}
-
 func BenchmarkArrOfPointers_32B_1_000(b *testing.B) {
 	runAoP32B(b, 1000)
 }
@@ -115,10 +121,6 @@ func BenchmarkArrOfPointers_32B_10_000(b *testing.B) {
 
 func BenchmarkArrOfPointers_32B_100_000(b *testing.B) {
 	runAoP32B(b, 100000)
-}
-
-func BenchmarkArrOfPointers_32B_250_000(b *testing.B) {
-	runAoP32B(b, 250000)
 }
 
 func BenchmarkArrOfPointers_64B_1_000(b *testing.B) {
@@ -133,10 +135,6 @@ func BenchmarkArrOfPointers_64B_100_000(b *testing.B) {
 	runAoP64B(b, 100000)
 }
 
-func BenchmarkArrOfPointers_64B_250_000(b *testing.B) {
-	runAoP64B(b, 250000)
-}
-
 func BenchmarkArrOfPointers_128B_1_000(b *testing.B) {
 	runAoP128B(b, 1000)
 }
@@ -149,10 +147,6 @@ func BenchmarkArrOfPointers_128B_100_000(b *testing.B) {
 	runAoP128B(b, 100000)
 }
 
-func BenchmarkArrOfPointers_128B_250_000(b *testing.B) {
-	runAoP128B(b, 250000)
-}
-
 func BenchmarkArrOfPointers_256B_1_000(b *testing.B) {
 	runAoP256B(b, 1000)
 }
@@ -163,8 +157,4 @@ func BenchmarkArrOfPointers_256B_10_000(b *testing.B) {
 
 func BenchmarkArrOfPointers_256B_100_000(b *testing.B) {
 	runAoP256B(b, 100000)
-}
-
-func BenchmarkArrOfPointers_256B_250_000(b *testing.B) {
-	runAoP256B(b, 250000)
 }